requirements.txt
werkzeug==2.2.2
flask==2.2.2
discord.py==2.4.0
python-dotenv==1.0.0